feat(schedule): implement schedule rule creation and retrieval with overlap prevention
