+++ backend/src/routes/schedule.routes.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const {
   createScheduleRule,
   getScheduleRulesByMember,
-} = require("../controllers/scheduleController");
+} = require("../controllers/schedule.controller");
 
 router.post("/", createScheduleRule);
 router.get("/:memberId", getScheduleRulesByMember);
